Created a new component called PhotoView, which has the photo and the navigation arrows
