+++ src/components/Photo.tsx
@@ -1,12 +1,11 @@
 import React from "react"
-import { Link, useNavigate } from "react-router-dom"
-import Drawer from "./Drawer"
-import NavigationDrawer from "./NavigationDrawer"
-import LeftArrow from "../assets/LeftArrow"
-import RightArrow from "../assets/RightArrow"
-import { PhotoData, BaseProps } from "../data/types"
+import { useNavigate } from "react-router-dom"
 import { useState, useContext } from "react"
 import { DrawerContext } from "../contexts/Drawer"
+import Drawer from "./Drawer"
+import PhotoView from "./PhotoView"
+import NavigationDrawer from "./NavigationDrawer"
+import { PhotoData, BaseProps } from "../data/types"
 
 interface PhotoProps extends BaseProps {
   photo?: PhotoData
@@ -17,7 +16,6 @@ function Photo({ photo, albumId, isLoading }: PhotoProps) {
   const { isOpen, toggle } = useContext(DrawerContext)
   const [touchPosition, setTouchPosition] = useState(null)
   const navigate = useNavigate()
-  const baseUrl = process.env.REACT_APP_API_BASE_URL
 
   function handleTouchStart(e: any) {
     const touchDown = e.touches[0].clientX
@@ -43,20 +41,6 @@ function Photo({ photo, albumId, isLoading }: PhotoProps) {
     }
     setTouchPosition(null)
   }
-
-  function stopPropagation(e: React.MouseEvent<HTMLDivElement, MouseEvent>) {
-    e.stopPropagation()
-  }
-
-  let id, prev, next
-  if (photo) {
-    id = photo.id
-    prev = photo.prev
-    next = photo.next
-  }
-  // const { id, prev, next } = photo
-  const photoLink = `${baseUrl}/api/v1/photos/${id}/file`
-
   return (
     <div
       className="full-photo-container"
@@ -64,21 +48,7 @@ function Photo({ photo, albumId, isLoading }: PhotoProps) {
       onTouchStart={handleTouchStart}
       onTouchMove={handleTouchMove}
     >
-      {prev && (
-        <div onClick={(e) => stopPropagation(e)}>
-          <Link to={`/albums/${albumId}/photos/${prev.id}`}>
-            <LeftArrow />
-          </Link>
-        </div>
-      )}
-      <img className="full-photo" src={photoLink} />
-      {next && (
-        <div onClick={(e) => stopPropagation(e)}>
-          <Link to={`/albums/${albumId}/photos/${next.id}`}>
-            <RightArrow />
-          </Link>
-        </div>
-      )}
+      <PhotoView photo={photo} albumId={albumId} isLoading={isLoading} />
       <Drawer photo={photo} isLoading={isLoading} />
       <NavigationDrawer
         photo={photo}
